Fix Issue 1770 for 3.0  (#1776)
* [CI] adding action for clang compiler
* [C] Ingnoring postgres warnings
* [dijkstra] compiling with clang
* [common] compiling with clang
* [astar] compiles with clang
* [ksp] compiles with clang
* [tsp] compiles with clang
* [alpha_shape] compiles with clang
* [tspEuclidean] compiles with clang
* [witPoints] compiles with clang
* [bellman_ford] compiles with clang
* [chinese] compiles with clang
* [transitiveClosure] compiles with clang
* [breadthFirstSearch] compiles with clang
* [components] compiles with clang
* [maxflow] pgr_maxflow.hpp compiles with clang
* [maxflow] compiles with clang
* [lineGraph] compiles with clang
* [bdAstar] compiles with clang
* [bdDijkstra] compiles with clang
* [trsp] compiles with clang
* [pickDeliver] compiles with clang
* [dijkstra][docqueries] giving a new correct alternate result
* Removing a test of a function that does not exist on pgRouting
* removing internal query tests that were created when developing contraction and are not used
* [tsp][docqueries] Modification of rand() gives new results
* roll back changes that affect ksp
* [C++] moving pgr_messages.cpp to cpp_common
* roll back changes that affect trsp
* src/common/basePath_SSEC.cpp compiles with clang
* src/trsp/pgr_trspHandler.cpp compiles with clang
* [ci][pgtap] adding testing after clang compilation
* [build] rolling back build-extension-update-files1.pl
* [C++] adding some minimal changes from the comments of the PR
* [docqueries] updating full-version & doc-pgr_extractVertices.result
* [doc] updating news & release notes & some result files

+++ include/max_flow/pgr_maxflow.hpp
@@ -103,7 +103,7 @@ class PgrFlowGraph {
                  graph,
                  supersource,
                  supersink);
-         return get_edge_disjoint_paths(flow);
+         return get_edge_disjoint_paths(static_cast<size_t>(flow));
      }
 
      PgrFlowGraph(
@@ -122,7 +122,7 @@ class PgrFlowGraph {
      std::vector<pgr_flow_t> get_flow_edges() const;
 
      std::vector<General_path_element_t> get_edge_disjoint_paths(
-             int64_t flow);
+             size_t flow);
 
  private:
      V get_boost_vertex(int64_t id) const {
@@ -152,7 +152,7 @@ class PgrFlowGraph {
 
      void flow_dfs(
              V vertex,
-             int64_t path_id,
+             size_t path_id,
              std::vector<std::vector<int64_t> > &paths);
 
      /*
